Renamed classes VirtualFileSystem to VFSProvider VFSContainer to VirtualFileSystem
