+++ VisualC/HatchGameEngine.vcxproj.filters
@@ -111,6 +111,9 @@
     <ClCompile Include="..\source\engine\filesystem\vfs\VFSEntry.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\source\engine\filesystem\vfs\VFSProvider.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="..\source\engine\filesystem\vfs\VirtualFileSystem.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
