+++ include/Engine/Filesystem/VFS/VirtualFileSystem.h
@@ -1,49 +1,32 @@
 #ifndef ENGINE_FILESYSTEM_VFS_VIRTUALFILESYSTEM_H
 #define ENGINE_FILESYSTEM_VFS_VIRTUALFILESYSTEM_H
 
-#include <Engine/Filesystem/VFS/VFSEntry.h>
+#include <Engine/Filesystem/VFS/VFSProvider.h>
 
-#include <Engine/Includes/Standard.h>
-#include <Engine/IO/Stream.h>
+#define DEFAULT_MOUNT_POINT ""
 
-enum VFSType {
-	FILESYSTEM,
-	HATCH
+enum VFSMountStatus {
+	NOT_FOUND = -3,
+	COULD_NOT_MOUNT = -2,
+	COULD_NOT_UNMOUNT = -1,
+	UNMOUNTED = 0,
+	MOUNTED = 1
 };
-
-#define VFS_READABLE 1
-#define VFS_WRITABLE 2
-
-typedef std::unordered_map<std::string, VFSEntry*> VFSEntryMap;
 
 class VirtualFileSystem {
 private:
-	Uint16 Flags;
-	std::string MountPoint;
+	std::deque<VFSProvider*> LoadedVFS;
 
-protected:
-	bool Opened;
+	void Dispose();
 
 public:
-	VirtualFileSystem(const char *mountPoint, Uint16 flags);
-	virtual ~VirtualFileSystem();
+	~VirtualFileSystem();
 
-	bool IsOpen();
-	bool IsReadable();
-	bool IsWritable();
+	VFSMountStatus Mount(const char* filename, const char* mountPoint, VFSType type, Uint16 flags);
+	int NumMounted();
 
-	const char* GetMountPoint();
-
-	virtual void TransformFilename(const char* filename, char* dest, size_t destSize);
-	virtual bool HasFile(const char* filename);
-	virtual VFSEntry* FindFile(const char* filename);
-	virtual bool ReadFile(const char* filename, Uint8** out, size_t* size);
-	virtual bool PutFile(const char* filename, VFSEntry* entry);
-	virtual bool EraseFile(const char* filename);
-	virtual Stream* OpenReadStream(const char* filename);
-	virtual Stream* OpenWriteStream(const char* filename);
-	virtual bool CloseStream(Stream*);
-	virtual void Close();
+	bool LoadFile(const char* filename, Uint8** out, size_t* size);
+	bool FileExists(const char* filename);
 };
 
 #endif /* ENGINE_FILESYSTEM_VFS_VIRTUALFILESYSTEM_H */
